[REFACTOR]: DurationView를 별도의 뷰로 빼지 않고 뷰 컨트롤러에서 생성하도록 변경
